requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
filelock==3.0.12
Flask==1.1.1
Flask-Cors==3.0.8
gdown==3.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.18.0
opencv-python==4.1.2.30
Pillow==6.2.1
PySocks==1.7.1
PyYAML==5.2
requests==2.22.0
six==1.13.0
torch==1.3.1
torchvision==0.4.2
tqdm==4.41.1
urllib3==1.25.7
Werkzeug==0.16.0
